trainer-project/requirements.txt
numpy==1.21.0
pandas==1.3.0
torch==1.9.0
scikit-learn==0.24.2
PyYAML==5.4.1